UI: Add "Add Magnet" and "Move Magnet" buttons
This adds two new buttons to the control panel of the example-sim:  "Add a
Magnet" and "Move the Magnet".  The move button simply moves the original magnet
to a random location on the screen.  The add button creates another magnet on the
screen for the user to manipulate. If the reset button is pressed, all new magnets
will be removed and the original magnet is reset as usual.  The "Flip Polarity"
button only affects the original magnet.